projects: adxrs290-pmdz: use the example infrastructure
The Makefile and src tree already use the shared examples infrastructure
(EXAMPLE := iio_example + examples.mk), but builds.json still passed the
old IIO_EXAMPLE=/IIO_TRIGGER_EXAMPLE=/IIO_TIMER_TRIGGER_EXAMPLE= toggles,
which the Makefile no longer reads - so every entry silently built the
default iio_example. Switch the flags to EXAMPLE=<dir> matching the active
example directory (iio_example, iio_trigger_example,
iio_timer_trigger_example).

While here, fix iio_trigger_example, which had never actually been
compiled.
