+++ projects/adxrs290-pmdz/src/platform/maxim/main.c
@@ -33,17 +33,38 @@
 
 #include "parameters.h"
 #include "common_data.h"
+#include "no_os_error.h"
 
 extern int example_main();
 
 /***************************************************************************//**
- * @brief Main function execution for STM32 platform.
+ * @brief Main function execution for Maxim platform.
  *
  * @return ret - Result of the enabled examples execution.
 *******************************************************************************/
 int main()
 {
+	int ret = -EINVAL;
+
 	adxrs290_ip.spi_init = adxrs290_spi_ip;
+
+#ifdef NVIC_GPIO_IRQ
+	struct no_os_irq_ctrl_desc *nvic_desc;
+	struct no_os_irq_init_param nvic_ip = {
+		.platform_ops = &max_irq_ops,
+	};
+
+	/* Initialize GPIO IRQ controller */
+	ret = no_os_irq_ctrl_init(&nvic_desc, &nvic_ip);
+	if (ret)
+		return ret;
+
+	ret = no_os_irq_enable(nvic_desc, NVIC_GPIO_IRQ);
+	if (ret) {
+		no_os_irq_ctrl_remove(nvic_desc);
+		return ret;
+	}
+#endif
 
 	return example_main();
 }
